+++ backend/tinyfarm/src/main/java/com/api/tinyfarm/repository/UserRepository.java
@@ -8,9 +8,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     
     Optional<User> findByName(String name);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
     Optional<List<User>> findByLevel(Integer level);
 }
